Eliminate black bars around videos
There doesn’t seem to be an option for the Windows Media Player control
to fill the screen eliminating black bars.

So here is little hack to get it working.

Get the aspect ratio of the videos. Use that to calculate the target
width of the form necessary for the video to fill the height. Finally
adjust the position of the form to the left the required amount to
centre the video.

Hopefully there is a smarter way to do this. But if not…