Prevent screensaver attempting to download movies twice
This should resolve the InternalMove IOExceptions on multiple monitors.

I think the bug is caused by each form (i.e. each monitor) attempting
to download the same file. And then once the download is complete, one
of the forms tries to move the file but can’t because the other form is
still writing to it. Resulting in the IOException error.