Clear temp folder on start
This is to remove MOV files that partially downloaded.

Tried putting this in a destructor in the form class to clean up when
the screensaver closes, but it was a bit hit and miss whether the
destructor succeeded in clearing up or not. And then if people have
multiple monitors you can have multiple trying to run at the same time.

So clearing when the screensaver launches, before it has created any
forms, seemed the best solution.